apm0013173-one-controller_kp/orm/requirements.txt
asgiref==3.3.4
Django==3.2.19
django-filter==2.4.0
djangorestframework==3.12.4
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.4
typing-extensions==3.10.0.0
